+++ tests/trybuild/empty_par.rs
@@ -1,3 +1,0 @@
-use besedarium::*;
-
-type EmptyPar = tpar!(Http;);